+++ ke-legacy/utils/menus/populate-menu/test.js
@@ -79,7 +79,7 @@ describe('ChickenServices', () => {
         const menu = populateMenu(lang,140,true);
         console.log(menu);
         expect(typeof menu).toEqual('object');
-        expect(JSON.stringify(menu)).toMatch("{\"0\":\"1) Make a payment\\n2) Check balance\\n3) Training\\n4) View transaction history\\n5) FAW Pesticide Order\\n6) Solar\\n7) Insurance\\n77)Next page\",\"1\":\"44)Previous page\\n8)Contact Call center\\n9) Locate an OAF duka\\n\"}");
+        expect(JSON.stringify(menu)).toMatch('{"0":"1) Make a payment\\n2) Check balance\\n3) Training\\n4) View transaction history\\n5) Prepayment amount\\n6) FAW Pesticide Order\\n7) Solar\\n77)Next page","1":"44)Previous page\\n8)Insurance\\n9) Contact Call center\\n10) Locate an OAF duka\\n"}');
 
     });
     it('should return an object of only options that satisfy the date condition if the character is greater than 140',()=>{        
@@ -87,8 +87,7 @@ describe('ChickenServices', () => {
         const menu = populateMenu(lang,140,true);
         console.log(menu);
         expect(typeof menu).toEqual('object');
-        expect(JSON.stringify(menu)).toMatch("{\"0\":\"1) Make a payment\\n2) Check balance\\n3) Training\\n4) View transaction history\\n5) FAW Pesticide Order\\n6) Solar\\n7) Insurance\\n77)Next page\",\"1\":\"44)Previous page\\n8)Contact Call center\\n9) Locate an OAF duka\\n\"}");
-
+        expect(JSON.stringify(menu)).toMatch('{"0":"1) Make a payment\\n2) Check balance\\n3) Training\\n4) View transaction history\\n5) Prepayment amount\\n6) FAW Pesticide Order\\n7) Solar\\n77)Next page","1":"44)Previous page\\n8)Insurance\\n9) Contact Call center\\n10) Locate an OAF duka\\n"}');
     });
     it('should not show enroll/topUp if the site is locked according to the dates',()=>{
         const mockCursor = { next: jest.fn(), 
@@ -103,8 +102,7 @@ describe('ChickenServices', () => {
         state.vars.isGroupLeader = true;
         var populateMenu = require('./populateMenu');
         const menu = populateMenu(lang,140,true);
-        expect(JSON.stringify(menu)).toMatch('{"0":"1) Make a payment\\n2) Check balance\\n3) JiT Top up\\n4) Training\\n5) View transaction history\\n6) FAW Pesticide Order\\n7) Solar\\n8) Insurance\\n77)Next page","1":"44)Previous page\\n9)Contact Call center\\n10) Locate an OAF duka\\n"}');
-
+        expect(JSON.stringify(menu)).toMatch('{"0":"1) Make a payment\\n2) Check balance\\n3) JiT Top up\\n4) Training\\n5) View transaction history\\n6) Prepayment amount\\n7) FAW Pesticide Order\\n77)Next page","1":"44)Previous page\\n8)Solar\\n9) Insurance\\n10) Contact Call center\\n11) Locate an OAF duka\\n"}');
     });
 
 });